repo: reorganize into services structure
