+++ cli/go.mod
@@ -1,0 +1,5 @@
+module github.com/crdsdev/crds/cli
+
+go 1.13
+
+require github.com/spf13/cobra v0.0.5
