--- go.mod
@@ -1,11 +1,0 @@
-module github.com/hasheddan/crds
-
-go 1.13
-
-require (
-	github.com/spf13/cobra v0.0.5
-	golang.org/x/net v0.0.0-20190827160401-ba9fcec4b297 // indirect
-	k8s.io/apiextensions-apiserver v0.0.0-20191017190756-60deea82a242
-	k8s.io/apimachinery v0.0.0-20191017185446-6e68a40eebf9
-	sigs.k8s.io/yaml v1.1.0
-)
