+++ internal/acronym/acronym.go
