+++ internal/go.mod
@@ -1,0 +1,9 @@
+module github.com/crdsdev/crds/internal
+
+go 1.13
+
+require (
+	k8s.io/apiextensions-apiserver v0.17.0
+	k8s.io/apimachinery v0.17.0
+	sigs.k8s.io/yaml v1.1.0
+)
