+++ internal/static/static.go
