+++ internal/static/static_test.go
