+++ services/static/go.mod
@@ -1,0 +1,8 @@
+module github.com/crdsdev/crds/services/static
+
+go 1.13
+
+require (
+	github.com/hasheddan/crds v0.0.0-20191202041052-424c67204fb7 // indirect
+	github.com/spf13/cobra v0.0.5
+)
